Fix: Handle pause() calls on destroyed uploads
This patch also improves the behavior for resume().

net.Stream currently does not guarantee 'data' to stop emitting after
pausing a stream. This makes the current throttling function
problematic and prone to race conditions. This patch will take of it
for now, but a further patch for node itself will make things even
smoother.