Feature: incomingForm 'fileBegin' event
This event is useful if you need to buffer the uploaded file to disk, but
you are interested in streaming it to somewhere else while it is still
being uploaded.

You may need this if you are streaming the file to a destination that is
potentially slower at receiving it than the client that is uploading. By
buffering to disk you can ensure maximum upload speed for the client, while
having the flexibility to streaming the file on disk with a different speed.