Add ability to control maxKeys in the querystring parser.
As of 8a98c2f1d81cabb6594dc388789d60d2f3f67c09 the Node querystring parser
will only extract 1000 parameters by default. This change puts the default
back to unlimited (to retain the original formidable behaviour) while allowing
it to be set via the `maxFields` option.

At the moment this option only affects the QuerystringParser, but could be
extended to apply to all the parsers for which it makes sense.