Require the close boundary
This removes the handling to not require the close boundary on the
request. If parsing without the cloding boundary is desired, it can
be done by the using when using the events interface.

In addition, not requiring the close boundary means if the connection
actually did end unexpectedly, but happed right on the boundary, then
there wouldn't have been an error.

closes #87