(test) update
the prevoius edition worked well for `arrayName[key]` however it couldn't proccess `arrayName[]`. This edition, as far as I can imagine covers all possible scenarios and probbly all degrees of nestedness within in an incoming form data. However, there is an added layer of stringification and that's why I was forced to change adjust the tested values.